// src/models/entrada-mercancia/etiqueta-pallet-bobina.model.ts
export class EtiquetaPalletBobinaModel {
    ipAddress: string;
    modelData: ModelData;
    zplData:   string;
}

export class ModelData {
    codQr:        string;
    descProducto: string;
    cliente:      string;
    medidas:      string;
    pesoBruto:    string;
    pesoNeto:     string;
    cantBobinas:  string;
    correlativo:  string;
    ordenFab:     string;
    codProducto:  string;
    fecha:        string;
    sapCorrel:    string;
}